backend/backend/requirements.txt
fastapi==0.95.0
uvicorn==0.18.3
psycopg2-binary==2.9.3
bcrypt==4.0.1
python-dotenv==0.21.0
psycopg2